requirements.txt
discord.py
PyNaCl
youtube-dl
python-dotenv
asyncio